Move NetworkScopeHierarchy type to its own type file
