+++ ui/apps/platform/src/Containers/NetworkGraph/types/networkScopeHierarchy.ts
@@ -1,0 +1,11 @@
+import { SearchFilter } from 'types/search';
+
+export type NetworkScopeHierarchy = {
+    cluster: {
+        id: string;
+        name: string;
+    };
+    namespaces: string[];
+    deployments: string[];
+    remainingQuery: Omit<SearchFilter, 'Cluster' | 'Namespace' | 'Deployment'>;
+};
